Refactored DB initialization code to Model assembly, and write a powershell script to create the migrations
