+++ AlwaysEncryptedSample.Models/AuthDbContext.cs
@@ -1,20 +1,24 @@
 using System.Data.Entity;
-using AlwaysEncryptedSample.Models;
-using AlwaysEncryptedSample.Properties;
+using AlwaysEncryptedSample.Models.Properties;
 using Microsoft.AspNet.Identity.EntityFramework;
 
-namespace AlwaysEncryptedSample.Services
+namespace AlwaysEncryptedSample.Models
 {
     public class AuthDbContext : IdentityDbContext<ApplicationUser>
     {
-        public AuthDbContext()
-            : base("DefaultConnection", throwIfV1Schema: true)
+        public AuthDbContext(string nameOrConnectionString)
+            : base(nameOrConnectionString, throwIfV1Schema: true)
         {
         }
 
         public static AuthDbContext Create()
         {
-            return new AuthDbContext();
+            return Create("name=DefaultConnection");
+        }
+
+        public static AuthDbContext Create(string nameOrConnectionString)
+        {
+            return new AuthDbContext(nameOrConnectionString);
         }
 
         /// <remarks>We override this to set the schema.</remarks>
