+++ AlwaysEncryptedSample.Models/Properties/Resources.resx
@@ -117,25 +117,30 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
-  <data name="EncryptedColumnsSQL" xml:space="preserve">
-    <value>SELECT
-		QUOTENAME(DB_NAME()) + '.' + QUOTENAME(s.name) + '.' +
-		QUOTENAME(t.name) AS FQTN,
-		s.name AS [Schema],
-		t.name AS [Table],
-		c.name AS ColumnName,
-		ct.name AS ColumnType,
-		c.max_length AS Length,
-		c.collation_name AS Collation,
-		c.is_nullable AS Nullable,
-		c.encryption_type_desc as EncryptionType,
-		c.generated_always_type_desc AS GeneratedAlways
-	FROM 
-		sys.columns c
-		INNER JOIN sys.types ct ON c.system_type_id = ct.system_type_id
-		INNER JOIN sys.tables t ON c.object_id = t.object_id
-		INNER JOIN  sys.schemas s on s.schema_id = t.schema_id
-	--WHERE encryption_type IS NOT NULL;
-	ORDER BY s.name, t.name, c.column_id;</value>
+  <data name="Log4NetDDL" xml:space="preserve">
+    <value>IF OBJECT_ID('[Logging].[Log]') IS NULL
+BEGIN
+	CREATE TABLE [Logging].[Log] (
+		[Id] [int] IDENTITY (1, 1) NOT NULL
+			CONSTRAINT PK_Log PRIMARY KEY  CLUSTERED,
+		[Date] [datetime] NOT NULL,
+		[Thread] [varchar] (255) NOT NULL,
+		[Level] [varchar] (50) NOT NULL,
+		[Logger] [varchar] (255) NOT NULL,
+		[User] [nvarchar] (50) NULL,
+		[ClientIP]
+		    [nvarchar](45) -- Why 45? Because IPv6 Why Nvarchar? Because Log4Net
+		    COLLATE  Latin1_General_BIN2
+		    ENCRYPTED WITH (
+		        COLUMN_ENCRYPTION_KEY = LogColumnsKey,
+                ENCRYPTION_TYPE = DETERMINISTIC,
+                ALGORITHM = 'AEAD_AES_256_CBC_HMAC_SHA_256'
+            ) NULL,
+		[Message] [varchar] (4000) NOT NULL,
+		[Exception] [varchar] (2000) NULL
+	);
+
+	CREATE INDEX IX_Logging_Log_Date ON Logging.Log([Date]) INCLUDE (Message,Exception);
+END</value>
   </data>
 </root>
