+++ AlwaysEncryptedSample.Models/app.config
@@ -3,6 +3,9 @@
   <configSections>
     <!-- For more information on Entity Framework configuration, visit http://go.microsoft.com/fwlink/?LinkID=237468 -->
     <section name="entityFramework" type="System.Data.Entity.Internal.ConfigFile.EntityFrameworkSection, EntityFramework, Version=6.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" requirePermission="false" />
+    <sectionGroup name="applicationSettings" type="System.Configuration.ApplicationSettingsGroup, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" >
+      <section name="AlwaysEncryptedSample.Models.Properties.Settings" type="System.Configuration.ClientSettingsSection, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" requirePermission="false" />
+    </sectionGroup>
   </configSections>
   <runtime>
     <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
@@ -38,4 +41,14 @@
       <provider invariantName="System.Data.SqlClient" type="System.Data.Entity.SqlServer.SqlProviderServices, EntityFramework.SqlServer" />
     </providers>
   </entityFramework>
+  <applicationSettings>
+    <AlwaysEncryptedSample.Models.Properties.Settings>
+      <setting name="AuthenticationSchema" serializeAs="String">
+        <value>Authentication</value>
+      </setting>
+      <setting name="CreditCardSchema" serializeAs="String">
+        <value>Purchasing</value>
+      </setting>
+    </AlwaysEncryptedSample.Models.Properties.Settings>
+  </applicationSettings>
 </configuration>
