+++ AlwaysEncryptedSample.Tests/AlwaysEncryptedSample.Tests.csproj
@@ -9,8 +9,8 @@
     <ProjectGuid>{D06DDFB4-18FE-4ED4-86B3-F8CE3F455C74}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>AlwayEncryptedSample.Tests</RootNamespace>
-    <AssemblyName>AlwayEncryptedSample.Tests</AssemblyName>
+    <RootNamespace>AlwaysEncryptedSample</RootNamespace>
+    <AssemblyName>AlwaysEncryptedSample.Tests</AssemblyName>
     <TargetFrameworkVersion>v4.6.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <ProjectTypeGuids>{3AC096D0-A1C2-E12C-1390-A8335801FDAB};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
