+++ AlwaysEncryptedSample/App_Start/Startup.Auth.cs
@@ -5,7 +5,6 @@ using Microsoft.Owin;
 using Microsoft.Owin.Security.Cookies;
 using Owin;
 using AlwaysEncryptedSample.Models;
-using AlwaysEncryptedSample.Services;
 
 namespace AlwaysEncryptedSample
 {
