+++ AlwaysEncryptedSample/Web.config
@@ -3,7 +3,7 @@
   <configSections>
     <section name="entityFramework" type="System.Data.Entity.Internal.ConfigFile.EntityFrameworkSection, EntityFramework, Version=6.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" requirePermission="false" />
     <sectionGroup name="applicationSettings" type="System.Configuration.ApplicationSettingsGroup, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089">
-      <section name="AlwaysEncryptedSample.Properties.Settings" type="System.Configuration.ClientSettingsSection, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" requirePermission="false" />
+      <section name="AlwaysEncryptedSample.Models.Properties.Settings" type="System.Configuration.ClientSettingsSection, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" requirePermission="false" />
     </sectionGroup>
     <section name="log4net" type="log4net.Config.Log4NetConfigurationSectionHandler, log4net" />
   <!-- For more information on Entity Framework configuration, visit http://go.microsoft.com/fwlink/?LinkID=237468 --></configSections>
@@ -187,7 +187,7 @@
     </providers>
   </entityFramework>
   <applicationSettings>
-    <AlwaysEncryptedSample.Properties.Settings>
+    <AlwaysEncryptedSample.Models.Properties.Settings>
       <!-- This is the schema name for the ASP.NET authentication schemas -->
       <setting name="AuthenticationSchema" serializeAs="String">
         <value>Authentication</value>
@@ -196,6 +196,6 @@
       <setting name="CreditCardSchema" serializeAs="String">
         <value>Purchasing</value>
       </setting>
-    </AlwaysEncryptedSample.Properties.Settings>
+    </AlwaysEncryptedSample.Models.Properties.Settings>
   </applicationSettings>
 </configuration>
