+++ appveyor/init.sql
@@ -1,8 +1,8 @@
-CREATE DATABASE $(SQL_SERVER_DATABASE);
+CREATE DATABASE [$(SQL_SERVER_DATABASE)];
 GO
 CREATE LOGIN $(SQL_SERVER_USER) WITH PASSWORD = '$(SQL_SERVER_PASSWORD)';
 GO
-USE $(SQL_SERVER_DATABASE);
+USE [$(SQL_SERVER_DATABASE)];
 GO
 CREATE USER $(SQL_SERVER_USER);
 ALTER ROLE  db_owner
